ServiceLayer update to StoreOrderSystem

+++ src/main/java/com/storeorderingsystem/storeorderingsystem/repository/Bill.java
@@ -1,11 +1,10 @@
 package com.storeorderingsystem.storeorderingsystem.repository;
 
-import org.springframework.data.annotation.Id;
-
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
 import jakarta.persistence.Table;
 
 @Entity
@@ -30,10 +29,10 @@ public class Bill {
 	private float totalPrice;
 	
 	@Column(name = "DISCOUNT_AMOUNT")
-    private int discountAmount; 
+    private double discountAmount; 
     
 	@Column(name = "BILL_AMOUNT")
-    private float billAmount;
+    private double billAmount;
 	
 	public String getBillId() {
 		return billId;
@@ -51,11 +50,11 @@ public class Bill {
 		this.createdDate = createdDate;
 	}
 
-	public float getBillAmount() {
+	public double getBillAmount() {
 		return billAmount;
 	}
 
-	public void setBillAmount(float billAmount) {
+	public void setBillAmount(double billAmount) {
 		this.billAmount = billAmount;
 	}
 
@@ -75,11 +74,11 @@ public class Bill {
 		this.itemsQuantityId = itemsQuantityId;
 	}
 
-	public int getDiscountAmount() {
+	public double getDiscountAmount() {
 		return discountAmount;
 	}
 
-	public void setDiscountAmount(int discountAmount) {
+	public void setDiscountAmount(double discountAmount) {
 		this.discountAmount = discountAmount;
 	}
